POC: Sketch of localfilesystem and commiting

+++ fsspec/spec.py
@@ -1,5 +1,15 @@
-
+from contextlib import contextmanager
 from .utils import read_block
+
+
+aliases = [
+    ('makedir', 'mkdir'),
+    ('listdir', 'ls'),
+    ('cp', 'copy'),
+    ('move', 'mv'),
+    ('rename', 'mv'),
+    ('delete', 'rm'),
+]
 
 
 class AbstractFileSystem(object):
@@ -16,7 +26,12 @@ class AbstractFileSystem(object):
 
         A reasonable default should be provided if there are no arguments
         """
+        self.autocommit = True
+        self._intrans = False
+        self._transaction = Transaction(self)
         self._singleton[0] = self
+        for new, old in aliases:
+            setattr(self, new, getattr(self, old))
 
     @classmethod
     def current(cls):
@@ -28,6 +43,15 @@ class AbstractFileSystem(object):
             return cls()
         else:
             return cls._singleton[0]
+
+    @property
+    def transaction(self):
+        """A context within which files are committed together upon exit
+
+        Requires the file class to implement `.commit()` and `.discard()`
+        for the normal and exception cases.
+        """
+        return self._transaction
 
     def invalidate_cache(self, path=None):
         """
@@ -56,6 +80,21 @@ class AbstractFileSystem(object):
             may be permissions, etc.
         """
         pass
+
+    def makedirs(self, path, exist_ok=False):
+        """Recursively make directories
+
+        Creates directory at path and any intervening required directories.
+        Raises exception if, for instance, the path already exists but is a
+        file.
+
+        Parameters
+        ----------
+        path: str
+            leaf directory name
+        exist_ok: bool (False)
+            If True, will error if the target already exists
+        """
 
     def rmdir(self, path):
         """Remove a directory, if empty"""
@@ -87,14 +126,22 @@ class AbstractFileSystem(object):
         """
         pass
 
-    def walk(self, path, detail=False):
+    def walk(self, path, simple=False):
         """ Return all files belows path
 
-        Like ``ls``, but recursing into subdirectories. If detail is False,
-        returns a list of full paths.
+        Similar to ``ls``, but recursing into subdirectories.
+
+        Parameters
+        ----------
+        path: str
+            Root to recurse into
+        simple: bool (False)
+            If True, returns a list of filenames. If False, returns an
+            iterator over tuples like
+            (dirpath, dirnames, filenames), see ``os.walk``.
         """
 
-    def du(self, path, total=False, deep=False):
+    def du(self, path, total=False):
         """Space used by files within a path
 
         If total is True, returns a number (bytes), if False, returns a
@@ -103,14 +150,12 @@ class AbstractFileSystem(object):
         Parameters
         ----------
         total: bool
-            whether to sum all the file sized
-        deep: bool
-            whether to descend into subdirectories.
+            whether to sum all the file sizes
         """
-        if deep:
-            sizes = {f['name']: f['size'] for f in self.walk(path, True)}
-        else:
-            sizes = {f['name']: f['size'] for f in self.ls(path, True)}
+        sizes = {}
+        for f in self.walk(path, True):
+            info = self.info(f)
+            sizes[info['name']] = info['size']
         if total:
             return sum(sizes.values())
         else:
@@ -147,20 +192,35 @@ class AbstractFileSystem(object):
 
     def exists(self, path):
         """Is there a file at the given path"""
-        pass
+        try:
+            self.info(path)
+            return True
+        except:
+            return False
 
     def info(self, path):
         """Give details of entry at path
 
         Returns a single dictionary, with exactly the same information as ``ls``
         would with ``detail=True``
+
+        Returns
+        -------
+        dict with keys: name (full path in the FS), size (in bytes), type (file,
+        directory, or something else) and other FS-specific keys.
         """
 
     def isdir(self, path):
         """Is this entry directory-like?"""
+        return self.info(path)['type'] == 'directory'
+
+    def isfile(self, path):
+        """Is this entry file-like?"""
+        return self.info(path)['type'] == 'file'
 
     def cat(self, path):
         """ Get the content of a file """
+        return self.open(path, 'rb').read()
 
     def get(self, rpath, lpath, **kwargs):
         """ Copy file to local
@@ -185,6 +245,7 @@ class AbstractFileSystem(object):
     def copy(self, path1, path2, **kwargs):
         """ Copy within two locations in the filesystem"""
 
+
     def mv(self, path1, path2, **kwargs):
         """ Move file from one location to another """
         self.copy(path1, path2, **kwargs)
@@ -202,6 +263,10 @@ class AbstractFileSystem(object):
             also remove the directory
         """
 
+    def _open(self, path, mode='rb', block_size=None, autocommit=True,
+              **kwargs):
+        pass
+
     def open(self, path, mode='rb', block_size=None, **kwargs):
         """
         Return a file-like object from the filesystem
@@ -211,6 +276,8 @@ class AbstractFileSystem(object):
 
         Parameters
         ----------
+        path: str
+            Target file
         mode: str like 'rb', 'w'
             See builtin ``open()``
         block_size: int
@@ -220,12 +287,24 @@ class AbstractFileSystem(object):
         if 'b' not in mode:
             mode = mode.replace('t', '') + 'b'
             return io.TextIOWrapper(
-                self.open(self, path, mode, block_size, **kwargs))
+                self.open(path, mode, block_size, **kwargs))
+        else:
+            ac = kwargs.pop('autocommit', not self._intrans)
+            if not self._intrans and not ac:
+                raise ValueError('Must use autocommit outside a transaction.')
+            f = self._open(path, mode=mode, block_size=block_size,
+                           autocommit=ac, **kwargs)
+            if not ac:
+                self.transaction.files.append(f)
+            return f
 
     def touch(self, path, **kwargs):
-        """ Create empty file """
-        with self.open(path, 'wb', **kwargs):
-            pass
+        """ Create empty file, or update timestamp """
+        if not self.exists(path):
+            with self.open(path, 'wb', **kwargs):
+                pass
+        else:
+            raise NotImplementedError  # update timestamp, if possible
 
     def read_block(self, fn, offset, length, delimiter=None):
         """ Read a block of bytes from
@@ -280,3 +359,32 @@ class AbstractFileSystem(object):
 
     def __setstate__(self, state):
         self.__dict__.update(state)
+
+
+class Transaction(object):
+    """Filesystem transaction context
+
+    Gathers files for deferred commit or discard, so that several write
+    operations can be finalized semi-atomically.
+    """
+
+    def __init__(self, fs):
+        self.fs = fs
+
+    def __enter__(self):
+        self.files = []
+        self.fs._intrans = True
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        # only commit if there was no exception
+        self.complete(commit=exc_type is None)
+
+    def complete(self, commit=True):
+        # TODO: define behaviour in case of exception during completion
+        for f in self.files:
+            if commit:
+                f.commit()
+            else:
+                f.discard()
+        self.files = []
+        self.fs._intrans = False
